+++ test/commands/clone_test.rb
@@ -7,6 +7,12 @@ class CloneTest < Test::Unit::TestCase
     assert_command input, command
   end
 
+  def test_private_clone_noop
+    input   = "--noop clone -p rtomayko/ronn"
+    command = "git clone git@github.com:rtomayko/ronn.git\n"
+    assert_output command, hub(input)
+  end
+
   def test_https_clone
     stub_https_is_preferred
     input   = "clone rtomayko/ronn"
@@ -26,15 +32,21 @@ class CloneTest < Test::Unit::TestCase
     assert_command input, command
   end
 
-  def test_your_public_clone
+  def test_your_clone_is_always_private
     input   = "clone resque"
-    command = "git clone git://github.com/tpw/resque.git"
+    command = "git clone git@github.com:tpw/resque.git"
+    assert_command input, command
+  end
+
+  def test_clone_repo_with_period
+    input   = "clone hookio/hook.js"
+    command = "git clone git://github.com/hookio/hook.js.git"
     assert_command input, command
   end
 
   def test_clone_with_arguments
     input   = "clone --bare -o master resque"
-    command = "git clone --bare -o master git://github.com/tpw/resque.git"
+    command = "git clone --bare -o master git@github.com:tpw/resque.git"
     assert_command input, command
   end
 
@@ -47,7 +59,7 @@ class CloneTest < Test::Unit::TestCase
       stub_github_user(nil)
     end
 
-    assert_equal "** No GitHub user set. See http://help.github.com/git-email-settings/\n", out
+    assert_equal "** No GitHub user set. See http://help.github.com/set-your-user-name-email-and-github-token/\n", out
   end
 
   def test_your_public_clone_fails_without_config
@@ -55,7 +67,7 @@ class CloneTest < Test::Unit::TestCase
       stub_github_user(nil)
     end
 
-    assert_equal "** No GitHub user set. See http://help.github.com/git-email-settings/\n", out
+    assert_equal "** No GitHub user set. See http://help.github.com/set-your-user-name-email-and-github-token/\n", out
   end
 
   def test_private_clone_left_alone
@@ -72,6 +84,10 @@ class CloneTest < Test::Unit::TestCase
 
   def test_normal_clone_from_path
     assert_forwarded "clone ./test"
+  end
+
+  def test_clone_with_host_alias
+    assert_forwarded "clone server:git/repo.git"
   end
 
   def test_alias_expand